[2441] Extract semantic creation from createAllocateEdge
Bug: https://github.com/eclipse-syson/syson/issues/2441
